fix polling
